atualizaçao funaço deletar mensagens

+++ commands/deletarMessagens.js
@@ -1,4 +1,6 @@
-const { SlashCommandBuilder } = require("discord.js")
+const { SlashCommandBuilder, EmbedBuilder, PermissionsBitField } = require("discord.js")
+
+
 
 //      Exportando e criando slashcommand
 module.exports = {
@@ -14,21 +16,37 @@ module.exports = {
     async execute(interaction) {
         //      Interaçao com com o slashcommand
         const quantidade = interaction.options.getInteger('quantidade');
+        const member = interaction.member;
 
-        if (quantidade <= 0) {
-            return interaction.reply("Você precisa colocar um número valido!")
-        }
-        //      Tratamento de erro devido a APi nao deixar excluir mensagem de +14 dias atras
-        try {
+        if (member.permissions.has(PermissionsBitField.Flags.Administrator)) 
+        {
 
-            const fetched = await interaction.channel.messages.fetch({ limit: quantidade });
+            if (quantidade <= 0) {
+                return interaction.reply("Você precisa colocar um número valido!")
+            }
+            //      Tratamento de erro devido a APi nao deixar excluir mensagem de +14 dias atras
+            try {
 
-            await interaction.channel.bulkDelete(fetched);
-            await interaction.reply(`Deletado ${quantidade} mensagens.`);
-        } catch (error) {
+                const fetched = await interaction.channel.messages.fetch({ limit: quantidade });
 
-            console.error(error);
-            await interaction.reply(`Não foi possível deletar mensagens devido a: ${error}`);
-        }
+                await interaction.channel.bulkDelete(fetched);
+                
+                //      Criaçao Embed
+                const embedInfotBot = new EmbedBuilder()
+                    .setColor(0x0099FF)
+                    .setAuthor({ name: 'ChiperBot', url: 'https://i.postimg.cc/yYXtx2X0/Whats-App-Image-2024-03-11-at-7-20-53-PM.jpg' })
+                    .setDescription(`O canal de texto ${interaction.channel} teve ${quantidade} mensagens deletadas por ${interaction.user.username}`)
+
+                await interaction.reply({ embeds: [embedInfotBot] });
+
+            } catch (error) {
+
+                console.error(error);
+                await interaction.reply(`Não foi possível deletar mensagens devido a: ${error}`);
+            }
+    } else {
+
+        await interaction.reply("Você nao tem permissão para utilizar este comando!")
     }
+  }
 }
